fix: Convert PositionFloat from decimal to double to resolve SQLite ordering issue (#85)
- Changed PlaylistTrack.PositionFloat from decimal to double type
- Removed precision configuration from database context
- Updated all DTOs and request models to use double
- Fixed all tests to use double literals instead of decimal
- Created migration to update database schema
- Updated API documentation to reflect type change

This resolves issue #84 where SQLite was unable to properly order by decimal
values in ORDER BY clauses. The double type is properly supported by SQLite
and maintains the fractional positioning system for conflict-free reordering.

+++ src/Audiarr.Api/Endpoints/PlaylistEndpoints.cs
@@ -180,7 +180,7 @@ public static class PlaylistEndpoints
                     .Where(t => request.InitialTrackIds.Contains(t.Id))
                     .ToListAsync();
 
-                decimal position = 0;
+                double position = 0;
                 foreach (var trackId in request.InitialTrackIds)
                 {
                     if (tracks.Any(t => t.Id == trackId))
@@ -414,7 +414,7 @@ public static class PlaylistEndpoints
             var existingTrackIds = playlist.PlaylistTracks.Select(pt => pt.TrackId).ToHashSet();
 
             // Determine starting position
-            decimal startPosition;
+            double startPosition;
             if (request.Position.HasValue && request.Position.Value >= 0)
             {
                 // Insert at specific position
